# docs/api/openapi-spec.yaml
openapi: 3.1.0
info:
  title: TimeSync Scheduling API
  version: 1.0.0
  description: |
    Enterprise AI Workforce Scheduling Platform REST API.

    **Authentication:** All endpoints (except /auth/*) require a Bearer JWT token.

    **Multi-tenancy:** All resources are automatically scoped to the authenticated user's organization.

    **Rate limiting:** 500 requests/minute per API key.
servers:
  - url: https://api.timesync.io/v1
    description: Production
  - url: https://staging-api.timesync.io/v1
    description: Staging
  - url: http://localhost:4000/api/v1
    description: Local development
security:
  - BearerAuth: []
...
paths:
...
  /schedules/{id}/generate:
    post:
      tags: [Schedules]
      summary: Trigger AI schedule generation
      description: |
        Asynchronously generates an optimized schedule using the AI engine.
        Returns a job_id to poll for results.
        Use GET /schedules/jobs/{jobId} to check status.
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateRequest'
      responses:
        '202':
          description: Generation started
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      id: { type: string, format: uuid }
                      status: { type: string, enum: [queued] }
  /schedules/jobs/{jobId}:
    get:
      tags: [Schedules]
      summary: Get generation job status
      parameters:
        - name: jobId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Job status and solutions (when completed)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerationJobResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
...
    GenerateRequest:
      type: object
      required: [startDate, endDate]
      properties:
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
        departmentIds:
          type: array
          items: { type: string, format: uuid }
        locationIds:
          type: array
          items: { type: string, format: uuid }
        optimizationGoals:
          type: array
          items:
            type: string
            enum: [minimize_cost, maximize_coverage, maximize_fairness, minimize_overtime, maximize_preference]
          default: [maximize_coverage, minimize_cost, maximize_fairness]
        maxRuntime:
          type: integer
          minimum: 5
          maximum: 120
          default: 30
    GenerationJobResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            id: { type: string, format: uuid }
            status:
              type: string
              enum: [queued, running, completed, failed, cancelled]
            progress: { type: integer, minimum: 0, maximum: 100 }
            solutions:
              type: array
              items:
                $ref: '#/components/schemas/GenerationSolution'
            hardViolations:
              type: array
              items:
                $ref: '#/components/schemas/RuleViolation'
            durationMs: { type: integer }
            summary:
              type: object
              properties:
                totalShifts: { type: integer }
                assignedShifts: { type: integer }
                unfilledShifts: { type: integer }
                estimatedCost: { type: number }
                overtimeHours: { type: number }
    GenerationSolution:
      type: object
      properties:
        index: { type: integer }
        label: { type: string }
        scores:
          type: object
          properties:
            coverage: { type: number }
            cost: { type: number }
            fairness: { type: number }
            preference: { type: number }
            overall: { type: number }
        stats:
          type: object
          properties:
            totalCost: { type: number }
            overtimeHours: { type: number }
            unfilledShifts: { type: integer }
        assignments:
          type: array
          items:
            $ref: '#/components/schemas/ShiftAssignment'
    RuleViolation:
      type: object
      properties:
        ruleId: { type: string }
        ruleName: { type: string }
        severity: { type: string, enum: [error, warning] }
        message: { type: string }
        affectedEntities:
          type: array
          items:
            type: object
            properties:
              type: { type: string }
              id: { type: string }
              name: { type: string }
    ShiftAssignment:
      type: object
      properties:
        shiftId: { type: string, format: uuid }
        employeeId: { type: string, format: uuid }
        status: { type: string }
        assignedBy: { type: string, enum: [manual, ai, self_pickup] }
        aiScore: { type: number }
        aiExplanation:
          type: object
          properties:
            reasons:
              type: array
              items:
                type: object
                properties:
                  factor: { type: string }
                  detail: { type: string }
                  weight: {}
                  score: { type: number }
            alternativesConsidered:
              type: array
              items: { type: string }
            overallScore: { type: number }
...